requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cffi==1.16.0
charset-normalizer==3.3.0
discord-interactions==0.4.0
discord-py-interactions==4.4.1
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
yarl==1.9.2
